third_party/blink/web_tests/plugins/mouse-capture-inside-shadow-expected.txt

Plugin received event: MouseEnter
Plugin received event: MouseMove
Plugin received event: MouseDown
Plugin received event: MouseMove
Plugin received event: MouseUp
This tests that mousedown captures the mouse correctly for a plugin object and continues to send events even if the mouse moves outside of the bounds of the plugin. On a successful run, plugin will see mousedown, followed by a mousemove, followed by a mouseup.
Console
plugin.mousedown
plugin.mousemove
plugin.mouseup
